- fixed bug on usage of LONG type in queries - auto conversion of big Integers in Longs - auto conversion to long of date in queries when compare a date field with a number

+++ core/src/main/java/com/orientechnologies/orient/core/sql/OSQLHelper.java
@@ -169,7 +169,7 @@ public class OSQLHelper {
 		if (t == OType.INTEGER)
 			return Integer.parseInt(iValue);
 		else if (t == OType.LONG)
-			return Integer.parseInt(iValue);
+			return Long.parseLong(iValue);
 		else if (t == OType.FLOAT)
 			return Float.parseFloat(iValue);
 		else if (t == OType.SHORT)
